Moved away from typescript and added testing platform
