+++ README.md
@@ -4,3 +4,9 @@
 
 * node >= v7.2.0
 * npm >= v3.10.10
+
+## Usage
+
+* npm run start
+* npm run start:dev
+* npm run test:api
