+++ package.json
@@ -1,7 +1,9 @@
 {
   "name": "generaptr",
   "scripts": {
-    "start": "gulp start"
+    "start": "node src/server.js",
+    "start:dev": "./node_modules/.bin/nodemon src/server.js",
+    "test:api": "./node_modules/.bin/mocha tests/api"
   },
   "version": "1.0.0",
   "description": "Api generator",
@@ -9,22 +11,15 @@
   "repository": "https://bitbucket.com/cupsadarius/generatr",
   "author": "Darius Cupsa <cupsadarius@gmail.com>",
   "keywords": [
-    "typescript",
     "nodejs",
     "express"
   ],
   "devDependencies": {
-    "@types/body-parser": "0.0.33",
-    "@types/express": "^4.0.33",
-    "@types/morgan": "^1.7.32",
-    "@types/node": "^6.0.45",
     "del": "^2.2.2",
-    "gulp": "^3.9.1",
-    "gulp-nodemon": "^2.2.1",
-    "gulp-tslint": "^6.1.2",
-    "run-sequence": "^1.2.2",
-    "tslint": "^3.15.1",
-    "typescript": "^2.0.3"
+    "mocha": "^3.2.0",
+    "nodemon": "^1.11.0",
+    "should": "^11.2.0",
+    "supertest": "^3.0.0"
   },
   "dependencies": {
     "body-parser": "^1.15.2",
