+++ src/configs/params.js
@@ -1,3 +1,3 @@
-export default {
+module.exports = {
     APP_PORT: process.env.PORT || 3000,
 };
