+++ src/configs/router.js
@@ -1,0 +1,6 @@
+const router = require('express').Router();
+const indexController = require('../controllers/indexController.js');
+
+router.use('/', indexController);
+
+module.exports = router;
