+++ src/controllers/indexController.js
@@ -1,0 +1,7 @@
+const router = require('express').Router();
+
+router.get('/', (req, resp) => {
+  resp.json({message: 'Api entrypoint'});
+});
+
+module.exports = router;
