--- src/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2015",
-    "module": "commonjs",
-    "sourceMap": true,
-    "removeComments": true,
-    "noImplicitAny": true,
-    "preserveConstEnums": true,
-    "rootDir": ".",
-    "outDir": "../dist",
-    "noFallthroughCasesInSwitch": true,
-    "allowJs": false
-  }
-}
